Prevent que worker crashing and preventing overmind from booting
If PostgreSQL isn't yet available, the que worker will exit 1. Overmind
will notice and cancel the boot for all other processes. This is
undesired behaviour. There is no way to express dependencies in
Procfiles, so we make do with simple solutions. Anyway, this is intended
to run in development mode, so no worries about a sleep.