Increase security with privileges and row-level security
PostgreSQL offers privileges to manage per-role access to tables and
functions. Use that to control what each role can access: tables,
columns, sequences, functions.

In addition, PostgreSQL also offers row-level security, where each row
is tested against predicates to determine if the row is a candidate for
the current user. This allows us to control what each group can see,
without forcing the UI to set "WHERE group_name = ?" everywhere. This is
both more powerful and more error-prone than the simple WHERE clause.

Row-level security is more powerful as there is no way to bypass the
system. If you are authenticated as user X, then each row will be tested
against the predicates. Row-level security is also more error-prone, as
there are more ways to cut yourself.

One choice that is made in scoutges is to copy the pgrole to every
table, so that the predicates are really simple: current_user = pgrole.
This makes the predicates very easy to assert against, and do not have
any performance impact, as each row is read anyway.

It is to be noted that row-level security is not an authorization
scheme. Row-level security is probably not the place to implement checks
like "if this person is an admin, then writes to this column are OK".

One other point: the sign in/registration case is the most complex. The
incoming request may be done in one of two roles: anonymous or within
the authenticated group. For all other tables, the work will be much
simpler as the policies will only need to assert `current_user =
pgrole`.