Prevent security hole by always creating new group roles
Imagine an attacker that guesses a group's name in the DB. They could
register themselves, and be automatically added to a group where they
might not belong.

The solution, as described in the commit, is to force authenticated
users to invite new users, and through the invitation, copy the current
user's security credentials to the newly invited person.