Prevent security hole with surprising behaviour between RLS and views
Row-level security and views interact in a surprising way: the view
"runs" as its owner, not per the current user. This is surprising
because the api.users view was created as migrator, but run as "10eme"
(for example). All users were returned, even users from other groups.

This was unexpected behaviour, but after thinking, it makes a certain
kind of sense. Row-level security & views are well explained in this
article:

   https://www.cybertec-postgresql.com/en/postgresql-row-level-security-views-and-a-lot-of-magic/