Remove, client-side, the possibility of having null party addresses
PostgreSQL's array_agg function will return an array with a single
`null` element when it has no rows to aggregate. This is undesirable
client-side, as it causes decoding failure. I ended up implementing a
solution server-side, but as a learning exercise, I wanted to implement
a solution client-side. I followed the example that lydell proposed at:

https://discourse.elm-lang.org/t/how-to-filter-a-json-decoded-list-to-remove-impossible-elements/5645/2